+++ test/test_slimta_relay_smtp_client.py
@@ -223,7 +223,7 @@ class TestSmtpRelayClient(unittest.TestCase, MoxTestBase):
         client._connect()
         client._ehlo()
         client._deliver(result, env)
-        self.assertEqual([Reply('250', 'Ok')], result.get_nowait())
+        self.assertEqual({'rcpt@example.com': Reply('250', 'Ok')}, result.get_nowait())
 
     def test_deliver_badpipeline(self):
         result = AsyncResult()
@@ -315,7 +315,7 @@ class TestSmtpRelayClient(unittest.TestCase, MoxTestBase):
         client._connect()
         client._ehlo()
         client._deliver(result, env)
-        self.assertEqual([Reply('250', 'Ok')], result.get_nowait())
+        self.assertEqual({'rcpt@example.com': Reply('250', 'Ok')}, result.get_nowait())
 
     def test_deliver_conversion_failure(self):
         result = AsyncResult()
@@ -370,7 +370,7 @@ class TestSmtpRelayClient(unittest.TestCase, MoxTestBase):
         self.mox.ReplayAll()
         client = SmtpRelayClient('addr', queue, socket_creator=self._socket_creator, ehlo_as='test')
         client._run()
-        self.assertEqual([Reply('250', 'Ok')], result.get_nowait())
+        self.assertEqual({'rcpt@example.com': Reply('250', 'Ok')}, result.get_nowait())
 
     def test_run_multiple(self):
         result1 = AsyncResult()
@@ -399,8 +399,8 @@ class TestSmtpRelayClient(unittest.TestCase, MoxTestBase):
         self.mox.ReplayAll()
         client = SmtpRelayClient('addr', queue, socket_creator=self._socket_creator, ehlo_as='test', idle_timeout=0.0)
         client._run()
-        self.assertEqual([Reply('250', 'Ok')], result1.get_nowait())
-        self.assertEqual([Reply('250', 'Ok')], result2.get_nowait())
+        self.assertEqual({'rcpt1@example.com': Reply('250', 'Ok')}, result1.get_nowait())
+        self.assertEqual({'rcpt2@example.com': Reply('250', 'Ok')}, result2.get_nowait())
 
     def test_run_random_exception(self):
         result = AsyncResult()